+++ tests/core/test_broker.py
@@ -1,5 +1,8 @@
 import json
 
+from _pytest.monkeypatch import MonkeyPatch
+
+from unittest.mock import patch
 import rasa.core.brokers.utils as broker_utils
 from rasa.core.brokers.file_producer import FileProducer
 from rasa.core.brokers.kafka import KafkaProducer
@@ -26,6 +29,30 @@ def test_pika_broker_from_config():
     assert actual.host == "localhost"
     assert actual.username == "username"
     assert actual.queue == "queue"
+
+
+def test_pika_message_property_app_id(monkeypatch: MonkeyPatch):
+    rasa_environment = "some-test-environment"
+
+    monkeypatch.setenv("RASA_ENVIRONMENT", rasa_environment)
+
+    # patch pika producer init
+    with patch.object(PikaProducer, "__init__", lambda: None):
+        pika_producer = PikaProducer()
+
+    # noinspection PyProtectedMember
+    assert pika_producer._message_properties.app_id == rasa_environment
+
+
+def test_pika_message_property_unset_app_id(monkeypatch: MonkeyPatch):
+    monkeypatch.delenv("RASA_ENVIRONMENT", raising=False)
+
+    # patch pika producer so it doesn't run `_run_pika()`
+    with patch.object(PikaProducer, "__init__", lambda: None):
+        pika_producer = PikaProducer()
+
+    # noinspection PyProtectedMember
+    assert not pika_producer._message_properties.app_id
 
 
 def test_no_broker_in_config():